dev-requirements.txt
-r requirements.txt
pytest>=5.0
pytest-django>=4.5.2
pytest-cov
django-debug-toolbar
sphinx
pre-commit
tqdm